proxyMasks
A method to calculate proxyMasks to be used for stencil bit masking.

Also introduces a new `m_prevProxies` private member, which is used to determine for which tiles a proxyMask should be
applied. (Might need some tweaking here).

Idea is, a proxy tile should not be drawn "only" where its respective non-proxy tiles are already drawn!